+++ circuit_maintenance_parser/__init__.py
@@ -11,9 +11,11 @@ from .parsers.ntt import ParserNTT
 from .parsers.packetfabric import ParserPacketFabric
 from .parsers.telstra import ParserTelstra
 from .parsers.zayo import ParserZayo
+from .parsers.cogent import ParserCogent
 
 
 SUPPORTED_PROVIDER_PARSERS = (
+    ParserCogent,
     ParserEUNetworks,
     ParserLumen,
     ParserMegaport,